+++ src/main/java/com/healthsync/Patient/Repository/PatientRepo.java
@@ -12,5 +12,5 @@ public interface PatientRepo extends JpaRepository<Patient, Integer> {
 
     boolean existsByMobileNumber(long mobileNumber);
 
-    Optional<Patient> findByMobileNumber(int mobileNumber);
+    Optional<Patient> findByMobileNumber(long mobileNumber);
 }
